Moro: Clean-up and documentation (section 1)
First commit in the effort to clean-up and document the MORO robustness optimization notebook. This commit adds a general introduction and cleans up and documents section 1 (Generate random policies) of the notebook.

Also adds the code to save the moro_robust_policies.pickle with a list of robust policies, to see if